TinyURL first version
This project has three main aspects
1. it takes input a long URL and converts it to a short URL.
2. it takes a short URL and provides the original URL.
3. And a short URL can be added or removed to a blacklist.
Blacklisted URLs cannot be used to get the original URL until it is
removed from the blacklist.
We have designed a simple user interface which interacts with the
RESTful APIs in the backend to complete this project. We have covered
most of the requirements provided in the Hackerrank question.